requirements.txt
fastapi==0.109.0
motor==3.3.2
pydantic==2.5.3
pymongo==4.6.1
python-dotenv==1.0.1
uvicorn==0.27.0.post1
